chore: resolve ssr options
